Added useResize custom hook
